Fix dates on view monitoring station page
https://eaflood.atlassian.net/browse/WATER-4736

In our last release we shipped a new version of the view monitoring station page. It replaces the legacy one. However, it was quickly fed back that the dates we display for tagged licences don't look right.

We still have access to the legacy view, and when we compared what it was showing it confirmed we'd got it wrong.

Having done some digging the problems are

- in the legacy view if there is no status update date, it shows a blank. We instead have opted to display the `createdAt` date
- When you tag a licence, if the licence has a licence condition you appear to be able to select it. The `licence_version_purpose_condition` is then linked to the `licence_monitoring_station` record. When this is the case, the legacy page uses the abstraction dates from the linked `licence_version_purpose`. We don't, and at time of writing it is not clear what we do fall back to

This change fixes the issues with dates we are displaying.