Update due to table name change
As part of this overall change we've opted to rename `review_charge_elements_returns` to `review_charge_element_returns`. It might _be_ more correct as a many-to-many table that we pluralise both parts.

But in use we only go one way and working with the table and model the current name always jars as 'feeling' wrong.