Add back in arrow-body-style rule
When we ran this through chatgpt it suggested we should switch this off. Googling for the rule in connection with prettier suggests there is an incompatibility between the two, hence best to switch it off.

But having double checked, the incompatibility is specific to [eslint-config-prettier](https://github.com/prettier/eslint-config-prettier?tab=readme-ov-file#arrow-body-style-and-prefer-arrow-callback) and its fix functionality.

Lots of blog posts that cover setting up prettier + eslint include adding eslint-config-prettier as an additional package to deal with any issues between the two.

However, when you dig in it is because they are also telling you to add eslint with its recommended rules. eslint-config-prettier then steps in and switches off those that prettier will handle instead.

But if you are specifying the rules as we are, rather than using the recommended ones, there is no need to add eslint-config-prettier. In fact prettier, [recommends avoiding it if you can](https://prettier.io/docs/en/integrating-with-linters.html).

> When searching for both Prettier and your linter on the Internet you'll probably find more related projects. These are generally not recommended[.]

eslint-config-itself states this rule is safe to use if

> You don't use eslint-plugin-prettier. In other words, you run `eslint --fix` and `prettier --write` as separate steps.

So, we add it back in to ensure our existing convention is followed.