Add JSDoc plugin/linting back in
This is admittedly where we could be seen to "cross the streams". prettier doesn't format JSDoc comments. There is a plugin (https://www.npmjs.com/package/prettier-plugin-jsdoc) though its not official. When we checked it we found its options were limited compared to eslint-plugin-jsdoc. Critically, there is no option to tell it to ignore @private or _myPrivateMethod() methods.

It also doesn't have the same level of support as eslint-plugin-jsdoc. So, though it could be argued it is more concerned with 'style', we'll continue to use eslint-plugin-jsdoc to lint our JSDoc comments.