Add code coverage report to job summary
During a team meeting it was suggested there might be an easier way of tracking missing code coverage if we could see it in the GitHub as we made changes.

SonarCloud has the ability to add a summary to your PR. But it does this for each change which gets noisy, plus it is still only the summary. The then mentioned that users of Jest will often expect to see a job summary detailing the coverage.

So, a quick check of that i available in GitHub actions led to [Write LCOV Report to Job Summary](https://github.com/marketplace/actions/write-lcov-report-to-job-summary).

We think it does just the trick!