Added StandardJS applicable ESlint core rules
By not referencing StandardJS as an extension any more there are a number of StandardJS rules we'd be missing if we left the config as is. The problem is StandardJS concerns itself with a number of the same rules as prettier.

So, we first copied all the core ESlint rules from https://github.com/standard/eslint-config-standard into our config (the ones that don't depend on an eslint plugin).

We then ran https://github.com/prettier/eslint-config-prettier against them to highlight any issues. There were quite a number. We removed all that were flagged, though we are going to go back and revisit this output in the next few changes.

```
The following rules are enabled with config that might conflict with Prettier. See:
https://github.com/prettier/eslint-config-prettier#special-rules

- curly
- no-tabs

The following rules are enabled but cannot be automatically checked. See:
https://github.com/prettier/eslint-config-prettier#special-rules

- no-mixed-operators
- no-unexpected-multiline
- quotes
```

We'll also look at the rules dependent on plugins that eslint-config-standard would add.