Add code coverage report to job summary (#1478)
During a team meeting, it was suggested that tracking missing code coverage might be easier if we could see it in GitHub as we made changes.

SonarCloud can add a summary to your PR. But it does this for each change, which gets noisy, and it is still only the summary. Then, it was mentioned that users of Jest will often expect to see a job summary detailing the coverage.

So, a quick check of that is available in GitHub actions led to [Write LCOV Report to Job Summary](https://github.com/marketplace/actions/write-lcov-report-to-job-summary).

We think it does just the trick!