Add back in core rule 'curly'
https://eslint.org/docs/latest/rules/curly

> JavaScript allows the omission of curly braces when a block contains only one statement. However, it is considered by many to be best practice to never omit curly braces around blocks, even when they are optional, because it can lead to bugs and reduces code clarity.

StandardJS rules were enforcing curly only if the block, for example `if`, was multiline (`curly: ['error', 'multi-line'],`). With that configuration this would be valid

```javascript
if (cart.items && cart.items[0] && cart.items[0].quantity === 0) updateCart(cart);
```

But if Prettier considered the line to long it would amend it to this.

```javascript
if (cart.items && cart.items[0] && cart.items[0].quantity === 0)
  updateCart(cart);
```

That would then cause `curly` to fail.

However, it is no longer an issue to use the rule if you avoid `'multi-line'` or `'multi-or-nest'` as the option and stick with `'all'`. That works for us, as it is already an unwritten convention that we always wrap our blocks, even one-line `if`, in braces.

So, we add the rule back in but use the `'all'` option. Running **eslint-config-prettier** doesn't flag this as an issue any more.