Add back in core 'no-tabs' rule
- https://eslint.org/docs/latest/rules/no-tabs
- https://eslint.style/rules/default/no-tabs

Good old JavaScript. It made dealing with this rule fun!

So, the core rule `'no-tabs'` conflicts with prettier, because it allows you to indent using tabs or spaces. Using `'no-tabs': 'error'` (which is how eslint-config-standardjs applies it) would find fault with that if prettier added them in.

Because of our .editorconfig file Prettier knows not to use tab characters. But to ensure there is no conflict eslint-config-prettier recommends using `"no-tabs": ["error", {"allowIndentationTabs": true}]`.

The fly in the ointment, is that `'no-tabs'` as a rule has been deprecated from ESLint from version 8. So, we could add it back in as eslint-config-prettier recommends. But then we are purposefully adding back in something that will eventually disappear.

The reason it is deprecated is that all stylistic ESlint rules were moved to https://eslint.style/ when ESLint v9 (which we are on) was released.

We still don't want any tabs in our code so we want this rule. But we believe the more forward looking solution is to reference the plugin ESlint expects you to use, even if that does mean we have to add the stylistic plugin back in!