Add back in core 'quotes' rule
- https://eslint.style/rules/default/quotes
- https://eslint.org/docs/latest/rules/quotes

Very similar to `'no-tabs'`, this is a core rule that eslint-config-standard adds in that eslint-config-prettier flags as an issue. But eslint-config-standard is also referencing the deprecated version of the rule. It has been moved to https://eslint.style/ from v9.

For reference, what eslint-config-standard is trying to do is ensure you don't inadvertently declare a string using backticks that doesn't need them, for example `hello, world`. Quotes also covers how you escape quotes, for example, if like us you use single quotes you want to still allow `"Hello world's"`.

eslint-config-prettier's concern is that your prettier `"singleQuote": true` is consistent with your `'quotes'` rule.

Fortunately, ours is so we can add this back in, but we'll reference the non-deprecated version.