Purposefully omit remaining core rules
The last two core rules left to contend with were

- https://eslint.org/docs/latest/rules/no-mixed-operators
- https://eslint.org/docs/latest/rules/no-unexpected-multiline

eslint-config-prettier explains why these my conflict with what prettier does

- https://github.com/prettier/eslint-config-prettier?tab=readme-ov-file#no-mixed-operators-deprecated
- https://github.com/prettier/eslint-config-prettier?tab=readme-ov-file#no-unexpected-multiline

Essentially, they are both purely stylistic and there is no configuration of them that would prevent prettier from being an issue. So, we have opted to stick with the purpose of this PR; we let Prettier decide!