Refactor remove bill licence supplementary billing flags
https://eaflood.atlassian.net/jira/software/c/projects/WATER/boards/96

As part of our efforts to clean up the legacy code base and ensure that our existing functionality is working, we have been refactoring and migrating the logic for supplementary billing flags. This PR is focused on migrating the mechanism for flagging a licence when it is removed from a bill run in a ready status.