Add eslint-plugin-prettier & eslint-config-prettier
With our config in its current guise, linting is now a separate 'thing' to formatting. That is of course unless the code you've added is incorrectly formatted. In which case we need to lint it?? To check? Don't we?!

Some of confusion comes from Prettier's own docs. On the front page you see.

> If you use ESLint [we do for code-quality rules], install eslint-config-prettier to make ESLint and Prettier play nice with each other. It turns off all ESLint rules that are unnecessary or might conflict with Prettier.

Along with that is a link to "Integrating with Linters". In there you'll find

> When searching for both Prettier and your linter on the Internet you'll probably find more related projects. These are generally not recommended[.]

When we started this journey, we veered towards the second instruction. We focused on getting prettier setup and working as the primary tool. We then added ESLint in separately and treated them separately.

But in use (when working on this branch) there different ethos and approach started to jar. Prettier (IMHO) is assuming you are just going to have your chosen IDE auto-format stuff on save. It views red squiggles as noise and annoyance, so developers should type away and let the plugin clean up afterwards.

ESLint on the overhand believes you want to know there is an issue, so leans into highlighting issues in the IDE assuming you'll fix as you code. Auto-fix is available, but as not all issues are covered it needs to let you know as you code.

Step forward [eslint-plugin-prettier](https://github.com/prettier/eslint-plugin-prettier)!

> Runs Prettier as an ESLint rule and reports differences as individual ESLint issues

Including it in our setup means prettier issues are now highlighted in the IDE along with all other ESLint code-quality issues. This means those who align with the ESLint ethos (highlight and fix as you code) can follow that workflow.

But as we still depend on keeping the prettier config (eslint-plugin-prettier will automatically use it) along with having to install the prettier dependency, those who prefer to just have the IDE clean up on save are free to do so.

We can move back to a single lint task both locally and CI, whilst devs have the flexibility in how issues are resolved (manually, ESLint auto-fix, or prettier + IDE format on save).