Add quarterly returns submissions to return versions (#1482)
* Add quarterly returns submissions to return versions

https://eaflood.atlassian.net/browse/WATER-4707

As part of the quarterly returns submissions work, we need to add a new checkbox to the additional-submission-options page for the Quarterly returns submissions.

When the return version start date is >= April 01 2025 then this is considered a quarterly return.

If the return version is for a water company and the return version start date is for a quarterly return. Then the quarterly return should be defaulted to true. This logic has been determined by the business.