Save quarterly returns submissions option
https://eaflood.atlassian.net/browse/WATER-4707

As part of the quarterly returns submissions work, we need to persist the quarterly returns option that has been either defaulted or set by the user.

This change the quarterlyReturns key to the generated return version before it is persisted.

We will only persist the 'quarterlyReturns' if the return version start date is for quarterly return submissions period (>= 2025-04-01).