Fix outstanding JSDoc issues
https://github.com/DEFRA/water-abstraction-team/issues/125

In [Lint JSDOC's](https://github.com/DEFRA/water-abstraction-system/pull/1269) we added linting of our JSDoc comments. If we are going to do it, we might as well do it properly and consistently!

There were quite a few, so since then we have cleaned up certain areas of the codebase, whilst configuring the linter to only warn of us of issues.

With this final change we deal with the remaining issues so we can switch the linter to error when it encounters an issue.