Fix error jsdoc/lines-before-block
The problem is caused by a clash of rules. We want a blank line before a JSDoc block, as it ensures we have some spacing between methods.

But StandardJS has a rule that prevents unnecessary padding in blocks. When trying to document a constructor in a class you get a clash.

The only solution we have is to disable the JSDoc rule within the affected file.