Fix back in view licence contact details text
https://eaflood.atlassian.net/browse/WATER-4559

Spotted when taking a new member through some areas of the system that the 'back link' we use in the view licence contact details page we built is inconsistent with all our other pages, including view licence purposes which branches from the same view licence summary page.

This change fixes the inconsistency.