Add abstraction method to licence purpose page
https://eaflood.atlassian.net/browse/WATER-4705

During the process of migrating the [licence points page](https://github.com/DEFRA/water-abstraction-system/pull/1451), the original plan was to include the abstraction method for each licence point.
However, after further consideration, it was decided to refactor and include the abstraction method within the licence purpose page instead.

This PR adds the abstraction method to each purpose on the licence purpose page.