Fix back link for copy existing in rtn vers setup
https://eaflood.atlassian.net/browse/WATER-4283

Spotted when prepping for a discussion about return versions setup that going back from the 'Use previous requirements for returns' page in the return version setup journey results in a 404.

Basically, when we carried out a refactor to [Rename return-requirements route return-versions](https://github.com/DEFRA/water-abstraction-system/pull/1431) we didn't spot that the back link had been built differently than all the other pages in the journey.

So, it is still trying to go back to the defunct route. This changes fixes it.