Rename columns in licence monitoring station view
The first step is to try and clear up some of the confusion caused by the misnamed fields in the table.

What is currently held in `restriction_type` is how we are measuring the water at the monitoring station for that licence. This will either be 'flow' or 'level'.

We don't have the specific details, but think of it as 'flow' is the volume of water passing through the monitoring station. 'Level' means the monitoring station is measuring the river level and calculating the volume of water based on it.

When you tag a licence, `measure_type` is automatically determine by the threshold unit the user selects.

```javascript
// src/internal/modules/gauging-stations/lib/helpers.js:61
const deduceRestrictionTypeFromUnit = unit => {
  const flowUnits = ['Ml/d', 'm3/s', 'm3/d', 'l/s']
  if (flowUnits.includes(unit)) {
    return 'flow'
  }
  return 'level'
}
```

`alter_type` is holding what we see on the screen as 'Restriction type', i.e. when the monitoring station reaches a certain threshold the licence is expected to stop or reduce its water abstraction.

For clarity, `status` is serving two functions. The default state is 'resume'. If there is no `date_status_updated` then we don't display it. If there is, it becomes the `alert_type` i.e. the last alert sent to the licensee instructed was a warn, stop, reduce, or resume instruction.