Link lic. ver. purpose cond. 2 lic. monitoring stn
Next step is to add the missing link in our models between licence monitoring stations and licence version purpose conditions.

When you tag a licence, if it has any conditions you can opt one when creating the `licence_monitoring_station` record. This means it won't have abstraction period information recorded against it. Instead, we have to go get it from the licence version purpose via the selected licence version purpose condition.

This is the primary reason why the dates we are showing in the UI are currently wrong. We're not pulling them from the licence version purpose when we're supposed to.