Housekeeping - remove unnecessary defaults
The intent of the helpers is that they will create a new record using the _minimum_ of values needed. For example, if a field cannot be null the helper will ensure it gets set.

But they are not intended to set anything that is not required. These values should come from the tests that use the helpers.