Rebuild the presenter to support 'proper' table
Our initial version of the presenter (and to an extent the fetch service that feeds it) was built to mimic how the legacy service displays records in the page.

A licence can be tagged to a monitoring station more than once. A common example we see is where a licence has been tagged with a 'reduce' restriction, and then tagged again at a lower threshold with a 'stop' restriction.

This results in two licence monitoring station records. When we display them on the page, we are still expected to show two rows in the table, but only display the licence reference (and link) once.

How the legacy page built it, and unfortunately how we copied it, was to add one row per licence, and if a licence has multiple records, add multiple values within the row cell.

There are two downsides to this. The first is the added complexity needed in both the view and the presenter to group the licence monitoring station records by licence, so the view sees a one-to-one relationship between rows and licences.

The second, is that it breaks the page! You only have to reduce the screen size slightly and the 'inner row' values become disconnected. Values in one column no longer line up with the others.

Better, is to build the table properly and use the features that HTML and the GOV.UK provide us. So, in this new version of the presenter all grouping is gone. For each licence monitoring station returned, we add a row to the table. This means the values line up correctly whatever the screen size, which has the added bonus of correcting the accessibility issues of the legacy implementation.

But we are still able to only display the licence ref and link to its view once by using the `rowspan` property. It means we have to build each row within the view, and check we only add the first cell (that contains the reference), once per licence. But when you strip out all the previous the logic on balance it still results in a simpler view. Plus a much simpler presenter.

Other than that we make some tweaks to fix the elements mentioned previously

- we correct the confusion around measure, restriction, and alert type
- we fix the abstraction period that gets displayed
- we make use of the existing `sentenceCase()` helper to remove some logic
- null is a better representation of something being empty than ''