Update the view to build the table properly
This matches the changes we've made to the presenter. Ignoring where we have renamed a value, the key change is we are adding a row for every licence monitoring station returned. This stops the table 'flow' breaking when the page display is reduced. It also means the table is structured as expected so better supports screen readers.

In order to still only display the first cell once for licences with multiple records, we have had to add logic that builds each row cell by cell.