Remove measure column and make heading dynamic
The comment we've added has more detail. But in summary, the measure type column look a little odd always showing the same measure for all linked records. It also meant the abstraction period column was more likely to be flow over two lines, increasing the need to scroll.

We still think it would be useful to know what type of measure is being used. So, instead with been a bit 'clever' with the column heading. Wonder if anyone spots it!?