Stop including deleted records
Initial testing of this updated page in pre-prod highlighted that we still seemed to have a discrepancy in what we were showing. The details were now right, but we were showing more licences than the legacy page.

Double checking the legacy query we spotted it was filtering out deleted licence monitoring stations (they are soft deleted). We clearly hadn't spotted that in our initial version, or this fixed version, till now!