Ensure ordering of reqs created from copy existing
https://eaflood.atlassian.net/browse/WATER-4744

When running [regression tests](https://github.com/DEFRA/water-abstraction-acceptance-tests) for the latest service release, the `Submit returns requirement (internal) using abstraction data` test consistently failed.

When we looked into the return version setup journey, nothing was especially wrong. As expected, the test still generates two return requirements from the abstraction data. However, the test asserts that **Example 1** was the first return requirement displayed on the `/check`` page, followed by **Example 2**.

It was failing because this was no longer the case. When we double-checked the code, we confirmed that no sort was applied. Basically, previous runs of the regression tests were just lucky!

We [fixed it](https://github.com/DEFRA/water-abstraction-system/pull/1473) but it led us to double check how the copy from an existing return version worked, and we found the same problem.

This change updates `GenerateFromExistingRequirementsService` to sort its results before returning them to `SubmitExistingService`. We also carry out some housekeeping and move all the 'existing' modules into their own folder. There is enough of them it helps to make it clearer which modules support this functionality.