Prep for testing sorting
For copy existing we can take care of the sorting in the fetch service. So, to get ready for testing that we want to ensure that the order we add we seed the records is _not_ the order we want to retrieve them. That way we can test the fetch service is ordering them by site description.

So, the first step is to alter the seeder to set the first return requirement's site description to something that alphabetically, should come after the second return requirement we seed.

We then update the tests to with the new site descriptions. This just makes sure our prep is not the cause of any failing tests.