Stop returning command-only async calls
There are two scenarios where folks will see @Cruikshanks sweep in and ask for a change.

```javascript
// Scenario 1
async function go (someId) {
  const result = await _fetch(someId)

  return result
}

// Scenario 2
async function go (someId) {
  await _updateStatus(someId)
}
```

In both cases he'll ask folks to just immediately return the call.

```javascript
// Scenario 1
async function go (someId) {
  return _fetch(someId)
}

// Scenario 2
async function go (someId) {
  return _updateStatus(someId)
}
```

In a recent change where we were adding a job, we make a command-only async call as a batch to `Promise.all()`. The original change had `await ProcessLicence.go(licenceId)` so in swooped @Cruikshanks !

This was challenged, so we carried out a mini spike and did some Googling and came to the conclusion.

- We have it bang on for scenario 1; we should return immediately
- @Cruikshanks has it _so wrong_ for scenario 2 🤦

The spike concluded that technically there is no difference. Using `await` rather than `return` doesn't block anything and the resolution is the same.

The Googling backed up the challenge made: by declaring `return _doThing()` it makes it look like we intend to return something when we don't. The fact that when we do this we then have to add the following to the JSDoc's probably should have clued us in earlier.

> `@returns {Promise<object>} the promise returned is not intended to resolve to any particular value`

We found a blog post that put into 'bigger' words (!!) why this is bad.

> `await` should be used for commands to ensure void is returned. The reason why void is returned for command is to enforce [command-query separation](https://en.wikipedia.org/wiki/Command%E2%80%93query_separation). Do not mix updates with reads as that violates [single-responsibility principle](https://en.wikipedia.org/wiki/Single-responsibility_principle).
>
> [When to use return vs await vs return await vs nothing](https://blog.battlefy.com/when-to-use-return-vs-await-vs-return-await-vs-nothing)

So, this change is @Cruikshanks right all the instances where he got it wrong!