Replace local const with return await
Because of our googling we now k now the difference between `return`, `await`, and `return await`.

If we just used `return` then we are saying it is on the calling function to resolve any error thrown. That doesn't work for us because we want to handle the failed call.

We want to return the results, and handle the error, hence our belief this was a case where we needed to await the result, then return it.

Now thanks to various blog posts we've read, what we needed was `return await`!

- https://jakearchibald.com/2017/await-vs-return-vs-return-await/
- https://stackoverflow.com/questions/38708550/difference-between-return-await-promise-and-return-promise
- https://blog.battlefy.com/when-to-use-return-vs-await-vs-return-await-vs-nothing