Refactor cancel bill run handling
Recently, we realised that sometimes we were returning async calls when in fact we should have been `await` them (see [Stop returning command-only async calls](https://github.com/DEFRA/water-abstraction-system/pull/1499)).

When we applied that update, we made a change to the cancel bill run and send bill run processes to replace this

```javascript
  _cancelBillRun(billRun).catch((error) => {
    global.GlobalNotifier.omfg(error.message, { billRunId }, error)
  })
```

with this

```javascript
  await _cancelBillRun(billRun)
```

Much cleaner! And we could get rid of the fake delays in the unit tests. Bonus!

But when double checking things manually we spotted we'd broken these processes. The intent was that

- the user is made to wait until the bill run status is updated (cancelling/sending)
- then the request is returned so we don't block the UI (user redirected to `/bill-runs`)
- app begins cancelling the bill run in the background

We needed the old logic to support the controller awaiting the call to, for example, `SubmitCancelBillRunService` for the status change, but then letting `SubmitCancelBillRunService` return control whilst cancelling the bill run in the background.

So, this wasn't related to incorrectly using `return myAsyncCall()`.

Simplest fix is to put back the changes. But the previous code felt like a hack and the delays in the unit tests 'icky'. With hindsight what these fudges tell us is we're asking the `SubmitCancelBillRunService` to do too much. If we split its responsibilities we could get it working again without these fudges.

So, this change refactors the cancel bill run process ahead of PR #1499, so we can pretend we never broke cancel bill run!